+++ src/ui/widgets/title-frame.lua
@@ -11,6 +11,7 @@ local Widgets = Addon.UserInterface.Widgets
     points? = table[],
     width? = number,
     height? = number,
+    onUpdateTooltip? = function(self, tooltip) -> nil,
     titleText? = string,
     titleTemplate? = string,
     titleJustify? = "LEFT" | "RIGHT" | "CENTER"
@@ -19,28 +20,36 @@ local Widgets = Addon.UserInterface.Widgets
 function Widgets:TitleFrame(options)
   -- Defaults.
   options.frameType = "Frame"
+  options.titleText = Addon:IfNil(options.titleText, ADDON_NAME)
+  options.titleTemplate = Addon:IfNil(options.titleTemplate, "GameFontNormal")
+  options.titleJustify = Addon:IfNil(options.titleJustify, "CENTER")
+
+  local onUpdateTooltip = options.onUpdateTooltip
+  options.onUpdateTooltip = nil
 
   -- Base frame.
   local frame = self:Frame(options)
 
+  -- Title button.
+  frame.titleButton = self:Frame({
+    name = "$parent_TitleBackground",
+    frameType = "Button",
+    parent = frame,
+    points = { { "TOPLEFT" }, { "TOPRIGHT" } },
+    onUpdateTooltip = onUpdateTooltip
+  })
+  frame.titleButton:SetBackdropColor(Colors.DarkGrey:GetRGBA(0.75))
+  frame.titleButton:RegisterForClicks("LeftButtonUp", "RightButtonUp")
+
   -- Title text.
-  frame.title = frame:CreateFontString("$parent_Title", "ARTWORK", options.titleTemplate or "GameFontNormal")
-  frame.title:SetText(Colors.White(options.titleText or ADDON_NAME))
+  frame.title = frame.titleButton:CreateFontString("$parent_Title", "ARTWORK", options.titleTemplate)
+  frame.title:SetText(Colors.White(options.titleText))
+  frame.title:SetPoint("LEFT", self:Padding(), 0)
+  frame.title:SetPoint("RIGHT", -self:Padding(), 0)
+  frame.title:SetJustifyH(options.titleJustify)
 
-  if options.titleJustify == "LEFT" then
-    frame.title:SetPoint("TOPLEFT", self:Padding(), -self:Padding())
-  elseif options.titleJustify == "RIGHT" then
-    frame.title:SetPoint("TOPRIGHT", -self:Padding(), -self:Padding())
-  else
-    frame.title:SetPoint("TOP", 0, -self:Padding())
-  end
-
-  -- Title background.
-  local titleHeight = max(frame.title:GetHeight(), frame.title:GetStringHeight()) + self:Padding(2)
-  frame.titleBackground = frame:CreateTexture("$parent_TitleBackground", "BACKGROUND", nil, 7)
-  frame.titleBackground:SetColorTexture(Colors.DarkGrey:GetRGBA(0.75))
-  frame.titleBackground:SetPoint("TOPLEFT")
-  frame.titleBackground:SetPoint("BOTTOMRIGHT", frame, "TOPRIGHT", 0, -titleHeight)
+  frame.titleButton:SetFontString(frame.title)
+  frame.titleButton:SetHeight(frame.title:GetStringHeight() + self:Padding(2))
 
   return frame
 end